# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digimon API",
        "description": "Documentacion de la API wrapper de digimon",
        "license": {
            "name": "Apache 3.2.4",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://digimon-api-x8ni.onrender.com/",
            "description": "Digimon API"
        }
    ],
    "paths": {
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "register",
                "requestBody": {
                    "description": "Información de usuario.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parametros inválidos."
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login con credenciales existentes",
                "operationId": "login",
                "requestBody": {
                    "description": "User login information",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "required": [
                                            "email"
                                        ],
                                        "type": "string"
                                    },
                                    "password": {
                                        "required": [
                                            "password"
                                        ],
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario logeado correctamente"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAi......."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No se pudo iniciar sesión."
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout usuario",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT para verificación de usiaro",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "Bearer {token}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se ha cerrado sesión correctamente"
                    }
                }
            }
        },
        "/api/v1/digimons": {
            "get": {
                "tags": [
                    "Digimon"
                ],
                "summary": "Listado de digimons",
                "description": "Retorna toda la lista de digimons",
                "operationId": "getDigimons",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT para verificación de usiaro",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "Bearer {token}"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero de página.",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Elemento por página.",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna lista de digimons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "digimons": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Digimon"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datos no encontrada."
                    },
                    "400": {
                        "description": "Parametros inválidos."
                    },
                    "502": {
                        "description": "Ha ocurrido un error inesperado."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/digimons/{id}": {
            "get": {
                "tags": [
                    "Digimon"
                ],
                "summary": "Listado de digimons",
                "description": "Retorna digimon info",
                "operationId": "getDigimonById",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT para verificación de usiaro",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "Bearer {token}"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del digimon.",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna información del digimon",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DigimonInfo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datos no encontrada."
                    },
                    "502": {
                        "description": "Ha ocurrido un error inesperado."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "user",
                "description": "Class User",
                "properties": {
                    "name": {
                        "required": [
                            "name"
                        ],
                        "type": "string"
                    },
                    "email": {
                        "required": [
                            "email"
                        ],
                        "type": "string"
                    },
                    "password": {
                        "required": [
                            "password"
                        ],
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Digimon": {
                "title": "digimon",
                "description": "Class Digimon list",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DigimonInfo": {
                "title": "digimon_info",
                "description": "Class Digimon Info",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "href": {
                                    "type": "string"
                                },
                                "transparent": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "levels": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "number"
                                },
                                "level": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "types": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "number"
                                },
                                "type": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "attributes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "number"
                                },
                                "attribute": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "number"
                                },
                                "field": {
                                    "type": "string"
                                },
                                "image": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "title": "pagination",
                "description": "Class Digimon list",
                "properties": {
                    "currentPage": {
                        "type": "number"
                    },
                    "elementsOnPage": {
                        "type": "number"
                    },
                    "totalElements": {
                        "type": "number"
                    },
                    "totalPages": {
                        "type": "number"
                    },
                    "previousPage": {
                        "type": "number"
                    },
                    "nextPage": {
                        "type": "number"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "JWT Token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Auth para usuarios por api"
        },
        {
            "name": "Digimon",
            "description": "Digimon api wrapper"
        }
    ]
}